+++ .JuliaFormatter.toml
@@ -3,3 +3,4 @@ annotate_untyped_fields_with_any = false
 format_markdown = true
 format_docstrings = true
 separate_kwargs_with_semicolon = true
+join_lines_based_on_source = false
