+++ .github/workflows/CI.yml
@@ -24,7 +24,7 @@ jobs:
           - Aqua
         version:
           - '1'
-          - '~1.10.0-0'
+          - '1.10'
     steps:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v2
@@ -46,6 +46,7 @@ jobs:
           coverage: false
         env:
           GROUP: ${{ matrix.group }}
+          RETESTITEMS_NWORKERS: 4
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
         with:
