+++ docs/pages.jl
@@ -1,17 +1,12 @@
 pages = [
     "DiffEqFlux.jl: High Level Scientific Machine Learning (SciML) Pre-Built Architectures" => "index.md",
-    "Differential Equation Machine Learning Tutorials" => Any["examples/neural_ode.md",
-        "examples/GPUs.md",
-        "examples/mnist_neural_ode.md",
-        "examples/mnist_conv_neural_ode.md",
-        "examples/augmented_neural_ode.md",
-        "examples/neural_sde.md",
-        "examples/collocation.md",
-        "examples/normalizing_flows.md",
-        "examples/hamiltonian_nn.md",
-        "examples/tensor_layer.md",
-        "examples/multiple_shooting.md",
-        "examples//neural_ode_weather_forecast.md"],
+    "Differential Equation Machine Learning Tutorials" => Any[
+        "examples/neural_ode.md", "examples/GPUs.md",
+        "examples/mnist_neural_ode.md", "examples/mnist_conv_neural_ode.md",
+        "examples/augmented_neural_ode.md", "examples/neural_sde.md",
+        "examples/collocation.md", "examples/normalizing_flows.md",
+        "examples/hamiltonian_nn.md", "examples/tensor_layer.md",
+        "examples/multiple_shooting.md", "examples//neural_ode_weather_forecast.md"],
     "Layer APIs" => Any["Classical Basis Layers" => "layers/BasisLayers.md",
         "Tensor Product Layer" => "layers/TensorLayer.md",
         "Continuous Normalizing Flows Layer" => "layers/CNFLayer.md",
@@ -19,5 +14,4 @@ pages = [
         "Neural Differential Equation Layers" => "layers/NeuralDELayers.md",
         "Hamiltonian Neural Network Layer" => "layers/HamiltonianNN.md"],
     "Utility Function APIs" => Any["Smoothed Collocation" => "utilities/Collocation.md",
-        "Multiple Shooting Functionality" => "utilities/MultipleShooting.md"]
-]
+        "Multiple Shooting Functionality" => "utilities/MultipleShooting.md"]]
