+++ src/DiffEqFlux.jl
@@ -1,21 +1,36 @@
 module DiffEqFlux
 
-import PrecompileTools
+using PrecompileTools: @recompile_invalidations
 
-PrecompileTools.@recompile_invalidations begin
-    using ADTypes, ChainRulesCore, ComponentArrays, ConcreteStructs, Functors,
-          LinearAlgebra, Lux, LuxCore, Random, Reexport, SciMLBase, SciMLSensitivity
-
-    # AD Packages
-    using ForwardDiff, Tracker, Zygote
-
-    # FFJORD Specific
-    using Distributions, DistributionsAD
+@recompile_invalidations begin
+    using ADTypes: ADTypes, AutoForwardDiff, AutoZygote, AutoEnzyme
+    using ChainRulesCore: ChainRulesCore
+    using ComponentArrays: ComponentArray
+    using ConcreteStructs: @concrete
+    using Distributions: Distributions, ContinuousMultivariateDistribution, Distribution,
+                         logpdf
+    using DistributionsAD: DistributionsAD
+    using ForwardDiff: ForwardDiff
+    using Functors: Functors, fmap
+    using LinearAlgebra: LinearAlgebra, Diagonal, det, diagind, mul!
+    using Lux: Lux, Chain, Dense, StatefulLuxLayer
+    using LuxCore: LuxCore, AbstractExplicitLayer, AbstractExplicitContainerLayer
+    using Random: Random, AbstractRNG, randn!
+    using Reexport: @reexport
+    using SciMLBase: SciMLBase, DAEProblem, DDEFunction, DDEProblem, EnsembleProblem,
+                     ODEFunction, ODEProblem, ODESolution, SDEFunction, SDEProblem, remake,
+                     solve
+    using SciMLSensitivity: SciMLSensitivity, AdjointLSS, BacksolveAdjoint, EnzymeVJP,
+                            ForwardDiffOverAdjoint, ForwardDiffSensitivity, ForwardLSS,
+                            ForwardSensitivity, GaussAdjoint, InterpolatingAdjoint, NILSAS,
+                            NILSS, QuadratureAdjoint, ReverseDiffAdjoint, ReverseDiffVJP,
+                            SteadyStateAdjoint, TrackerAdjoint, TrackerVJP, ZygoteAdjoint,
+                            ZygoteVJP
+    using Tracker: Tracker
+    using Zygote: Zygote
 end
 
-import ChainRulesCore as CRC
-import LuxCore: AbstractExplicitLayer, AbstractExplicitContainerLayer
-import Lux.Experimental: StatefulLuxLayer
+const CRC = ChainRulesCore
 
 @reexport using ADTypes, Lux
 
@@ -31,8 +46,8 @@ export NeuralODE, NeuralDSDE, NeuralSDE, NeuralCDDE, NeuralDAE, AugmentedNDELaye
        NeuralODEMM, TensorLayer, SplineLayer
 export NeuralHamiltonianDE, HamiltonianNN
 export FFJORD, FFJORDDistribution
-export TensorProductBasisFunction,
-       ChebyshevBasis, SinBasis, CosBasis, FourierBasis, LegendreBasis, PolynomialBasis
+export TensorProductBasisFunction, ChebyshevBasis, SinBasis, CosBasis, FourierBasis,
+       LegendreBasis, PolynomialBasis
 export DimMover
 
 export EpanechnikovKernel, UniformKernel, TriangularKernel, QuarticKernel, TriweightKernel,
@@ -45,8 +60,8 @@ export multiple_shoot
 # Reexporting only certain functions from SciMLSensitivity
 export BacksolveAdjoint, QuadratureAdjoint, GaussAdjoint, InterpolatingAdjoint,
        TrackerAdjoint, ZygoteAdjoint, ReverseDiffAdjoint, ForwardSensitivity,
-       ForwardDiffSensitivity, ForwardDiffOverAdjoint, SteadyStateAdjoint,
-       ForwardLSS, AdjointLSS, NILSS, NILSAS
+       ForwardDiffSensitivity, ForwardDiffOverAdjoint, SteadyStateAdjoint, ForwardLSS,
+       AdjointLSS, NILSS, NILSAS
 export TrackerVJP, ZygoteVJP, EnzymeVJP, ReverseDiffVJP
 
 end
