+++ src/ffjord.jl
@@ -50,12 +50,12 @@ References:
 end
 
 function LuxCore.initialstates(rng::AbstractRNG, n::FFJORD)
-    return (;
-        model = LuxCore.initialstates(rng, n.model), regularize = false, monte_carlo = true)
+    return (; model = LuxCore.initialstates(rng, n.model),
+        regularize = false, monte_carlo = true)
 end
 
-function FFJORD(model, tspan, input_dims, args...; ad = AutoForwardDiff(),
-        basedist = nothing, kwargs...)
+function FFJORD(model, tspan, input_dims, args...;
+        ad = AutoForwardDiff(), basedist = nothing, kwargs...)
     !(model isa AbstractExplicitLayer) && (model = Lux.transform(model))
     return FFJORD(model, basedist, ad, input_dims, tspan, args, kwargs)
 end
@@ -68,8 +68,8 @@ function __jacobian_with_ps(model, psax, N, x)
     end
 end
 
-function __jacobian(::AutoForwardDiff{nothing}, model, x::AbstractMatrix,
-        ps::ComponentArray)
+function __jacobian(
+        ::AutoForwardDiff{nothing}, model, x::AbstractMatrix, ps::ComponentArray)
     psd = getdata(ps)
     psx = vcat(vec(x), psd)
     N = length(x)
@@ -123,8 +123,8 @@ end
 
 __norm_batched(x) = sqrt.(sum(abs2, x; dims = 1:(ndims(x) - 1)))
 
-function __ffjord(model, u, p, ad = AutoForwardDiff(), regularize::Bool = false,
-        monte_carlo::Bool = true)
+function __ffjord(model, u, p, ad = AutoForwardDiff(),
+        regularize::Bool = false, monte_carlo::Bool = true)
     N = ndims(u)
     L = size(u, N - 1)
     z = selectdim(u, N - 1, 1:(L - ifelse(regularize, 3, 1)))
@@ -141,8 +141,8 @@ function __ffjord(model, u, p, ad = AutoForwardDiff(), regularize::Bool = false,
         eJ = vec(e)' * reshape(J, size(J, 1), :)
     end
     if regularize
-        return cat(mz, -trace_jac, sum(abs2, mz; dims = 1:(N - 1)), __norm_batched(eJ);
-            dims = Val(N - 1))
+        return cat(mz, -trace_jac, sum(abs2, mz; dims = 1:(N - 1)),
+            __norm_batched(eJ); dims = Val(N - 1))
     else
         return cat(mz, -trace_jac; dims = Val(N - 1))
     end
@@ -155,17 +155,16 @@ function __forward_ffjord(n::FFJORD, x, ps, st)
     (; regularize, monte_carlo) = st
     sensealg = InterpolatingAdjoint(; autojacvec = ZygoteVJP())
 
-    model = StatefulLuxLayer(n.model, nothing, st.model)
+    model = StatefulLuxLayer{true}(n.model, nothing, st.model)
 
     ffjord(u, p, t) = __ffjord(model, u, p, n.ad, regularize, monte_carlo)
 
     _z = ChainRulesCore.@ignore_derivatives fill!(
-        similar(x,
-            S[1:(N - 2)]..., ifelse(regularize, 3, 1), S[N]), zero(T))
+        similar(x, S[1:(N - 2)]..., ifelse(regularize, 3, 1), S[N]), zero(T))
 
     prob = ODEProblem{false}(ffjord, cat(x, _z; dims = Val(N - 1)), n.tspan, ps)
-    sol = solve(prob, n.args...; sensealg, n.kwargs..., save_everystep = false,
-        save_start = false, save_end = true)
+    sol = solve(prob, n.args...; sensealg, n.kwargs...,
+        save_everystep = false, save_start = false, save_end = true)
     pred = __get_pred(sol)
     L = size(pred, N - 1)
 
@@ -215,17 +214,16 @@ function __backward_ffjord(::Type{T1}, n::FFJORD, n_samples::Int, ps, st, rng) w
     (; regularize, monte_carlo) = st
     sensealg = InterpolatingAdjoint(; autojacvec = ZygoteVJP())
 
-    model = StatefulLuxLayer(n.model, nothing, st.model)
+    model = StatefulLuxLayer{true}(n.model, nothing, st.model)
 
     ffjord(u, p, t) = __ffjord(model, u, p, n.ad, regularize, monte_carlo)
 
     _z = ChainRulesCore.@ignore_derivatives fill!(
-        similar(x,
-            S[1:(N - 2)]..., ifelse(regularize, 3, 1), S[N]), zero(T))
+        similar(x, S[1:(N - 2)]..., ifelse(regularize, 3, 1), S[N]), zero(T))
 
     prob = ODEProblem{false}(ffjord, cat(x, _z; dims = Val(N - 1)), reverse(n.tspan), ps)
-    sol = solve(prob, n.args...; sensealg, n.kwargs..., save_everystep = false,
-        save_start = false, save_end = true)
+    sol = solve(prob, n.args...; sensealg, n.kwargs...,
+        save_everystep = false, save_start = false, save_end = true)
     pred = __get_pred(sol)
     L = size(pred, N - 1)
 
@@ -268,8 +266,8 @@ end
 function Distributions._logpdf(d::FFJORDDistribution, x::AbstractArray)
     return first(first(__forward_ffjord(d.model, x, d.ps, d.st)))
 end
-function Distributions._rand!(rng::AbstractRNG, d::FFJORDDistribution,
-        x::AbstractArray{<:Real})
+function Distributions._rand!(
+        rng::AbstractRNG, d::FFJORDDistribution, x::AbstractArray{<:Real})
     x[:] = __backward_ffjord(eltype(d), d.model, size(x, ndims(x)), d.ps, d.st, rng)
     return x
 end
