+++ src/neural_de.jl
@@ -15,7 +15,8 @@ derivatives of the loss backwards in time.
 
 Arguments:
 
-  - `model`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the ̇x.
+  - `model`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
+    ̇x.
   - `tspan`: The timespan to be solved on.
   - `alg`: The algorithm used to solve the ODE. Defaults to `nothing`, i.e. the
     default algorithm from DifferentialEquations.jl.
@@ -44,7 +45,7 @@ function NeuralODE(model, tspan, args...; kwargs...)
 end
 
 function (n::NeuralODE)(x, p, st)
-    model = StatefulLuxLayer(n.model, nothing, st)
+    model = StatefulLuxLayer{true}(n.model, nothing, st)
 
     dudt(u, p, t) = model(u, p)
     ff = ODEFunction{false}(dudt; tgrad = basic_tgrad)
@@ -64,10 +65,10 @@ Constructs a neural stochastic differential equation (neural SDE) with diagonal 
 
 Arguments:
 
-  - `drift`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the drift
-    function.
-  - `diffusion`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
-    diffusion function. Should output a vector of the same size as the input.
+  - `drift`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
+    drift function.
+  - `diffusion`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines
+    the diffusion function. Should output a vector of the same size as the input.
   - `tspan`: The timespan to be solved on.
   - `alg`: The algorithm used to solve the ODE. Defaults to `nothing`, i.e. the
     default algorithm from DifferentialEquations.jl.
@@ -92,8 +93,8 @@ function NeuralDSDE(drift, diffusion, tspan, args...; kwargs...)
 end
 
 function (n::NeuralDSDE)(x, p, st)
-    drift = StatefulLuxLayer(n.drift, nothing, st.drift)
-    diffusion = StatefulLuxLayer(n.diffusion, nothing, st.diffusion)
+    drift = StatefulLuxLayer{true}(n.drift, nothing, st.drift)
+    diffusion = StatefulLuxLayer{true}(n.diffusion, nothing, st.diffusion)
 
     dudt(u, p, t) = drift(u, p.drift)
     g(u, p, t) = diffusion(u, p.diffusion)
@@ -106,16 +107,16 @@ end
 
 """
     NeuralSDE(drift, diffusion, tspan, nbrown, alg = nothing, args...;
-        sensealg=TrackerAdjoint(),kwargs...)
+        sensealg=TrackerAdjoint(), kwargs...)
 
 Constructs a neural stochastic differential equation (neural SDE).
 
 Arguments:
 
-  - `drift`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the drift
-    function.
-  - `diffusion`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
-    diffusion function. Should output a matrix that is `nbrown x size(x, 1)`.
+  - `drift`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
+    drift function.
+  - `diffusion`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines
+    the diffusion function. Should output a matrix that is `nbrown x size(x, 1)`.
   - `tspan`: The timespan to be solved on.
   - `nbrown`: The number of Brownian processes.
   - `alg`: The algorithm used to solve the ODE. Defaults to `nothing`, i.e. the
@@ -142,8 +143,8 @@ function NeuralSDE(drift, diffusion, tspan, nbrown, args...; kwargs...)
 end
 
 function (n::NeuralSDE)(x, p, st)
-    drift = StatefulLuxLayer(n.drift, p.drift, st.drift)
-    diffusion = StatefulLuxLayer(n.diffusion, p.diffusion, st.diffusion)
+    drift = StatefulLuxLayer{true}(n.drift, p.drift, st.drift)
+    diffusion = StatefulLuxLayer{true}(n.diffusion, p.diffusion, st.diffusion)
 
     dudt(u, p, t) = drift(u, p.drift)
     g(u, p, t) = diffusion(u, p.diffusion)
@@ -169,8 +170,8 @@ Arguments:
     and produce and output shaped like `x`.
   - `tspan`: The timespan to be solved on.
   - `hist`: Defines the history function `h(u, p, t)` for values before the start of the
-    integration. Note that `u` is supposed to be used to return a value that matches the size
-    of `u`.
+    integration. Note that `u` is supposed to be used to return a value that matches the
+    size of `u`.
   - `lags`: Defines the lagged values that should be utilized in the neural network.
   - `alg`: The algorithm used to solve the ODE. Defaults to `nothing`, i.e. the
     default algorithm from DifferentialEquations.jl.
@@ -195,7 +196,7 @@ function NeuralCDDE(model, tspan, hist, lags, args...; kwargs...)
 end
 
 function (n::NeuralCDDE)(x, ps, st)
-    model = StatefulLuxLayer(n.model, nothing, st)
+    model = StatefulLuxLayer{true}(n.model, nothing, st)
 
     function dudt(u, h, p, t)
         xs = mapfoldl(lag -> h(p, t - lag), vcat, n.lags)
@@ -203,8 +204,8 @@ function (n::NeuralCDDE)(x, ps, st)
     end
 
     ff = DDEFunction{false}(dudt; tgrad = basic_dde_tgrad)
-    prob = DDEProblem{false}(ff, x, (p, t) -> n.hist(x, p, t), n.tspan, ps;
-        constant_lags = n.lags)
+    prob = DDEProblem{false}(
+        ff, x, (p, t) -> n.hist(x, p, t), n.tspan, ps; constant_lags = n.lags)
 
     return (solve(prob, n.args...; sensealg = TrackerAdjoint(), n.kwargs...), model.st)
 end
@@ -240,15 +241,15 @@ Arguments:
     kwargs
 end
 
-function NeuralDAE(model, constraints_model, tspan, args...; differential_vars = nothing,
-        kwargs...)
+function NeuralDAE(
+        model, constraints_model, tspan, args...; differential_vars = nothing, kwargs...)
     !(model isa AbstractExplicitLayer) && (model = Lux.transform(model))
     return NeuralDAE(model, constraints_model, tspan, args, differential_vars, kwargs)
 end
 
 function (n::NeuralDAE)(u_du::Tuple, p, st)
     u0, du0 = u_du
-    model = StatefulLuxLayer(n.model, nothing, st)
+    model = StatefulLuxLayer{true}(n.model, nothing, st)
 
     function f(du, u, p, t)
         nn_out = model(vcat(u, du), p)
@@ -287,7 +288,8 @@ constraint equations.
 
 Arguments:
 
-  - `model`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the ̇`f(u,p,t)`
+  - `model`: A `Flux.Chain` or `Lux.AbstractExplicitLayer` neural network that defines the
+    ̇`f(u,p,t)`
   - `constraints_model`: A function `constraints_model(u,p,t)` for the fixed constraints to
     impose on the algebraic equations.
   - `tspan`: The timespan to be solved on.
@@ -320,7 +322,7 @@ function NeuralODEMM(model, constraints_model, tspan, mass_matrix, args...; kwar
 end
 
 function (n::NeuralODEMM)(x, ps, st)
-    model = StatefulLuxLayer(n.model, nothing, st)
+    model = StatefulLuxLayer{true}(n.model, nothing, st)
 
     function f(u, p, t)
         nn_out = model(u, p)
@@ -349,7 +351,9 @@ Arguments:
 
 References:
 
-[1] Dupont, Emilien, Arnaud Doucet, and Yee Whye Teh. "Augmented neural ODEs." In Proceedings of the 33rd International Conference on Neural Information Processing Systems, pp. 3140-3150. 2019.
+[1] Dupont, Emilien, Arnaud Doucet, and Yee Whye Teh. "Augmented neural ODEs." In
+Proceedings of the 33rd International Conference on Neural Information Processing
+Systems, pp. 3140-3150. 2019.
 """
 function AugmentedNDELayer(model::Union{NeuralDELayer, NeuralSDELayer}, adim::Int)
     return Chain(Base.Fix2(__augment, adim), model)
@@ -362,8 +366,7 @@ end
 
 function __augment(x::AbstractArray, augment_dim::Int)
     y = CRC.@ignore_derivatives fill!(
-        similar(x, size(x)[1:(ndims(x) - 2)]...,
-            augment_dim, size(x, ndims(x))), 0)
+        similar(x, size(x)[1:(ndims(x) - 2)]..., augment_dim, size(x, ndims(x))), 0)
     return cat(x, y; dims = Val(ndims(x) - 1))
 end
 
@@ -372,9 +375,9 @@ end
 
 Constructs a Dimension Mover Layer.
 
-We can have Flux's conventional order `(data, channel, batch)` by using it as the last layer
-of `Flux.Chain` to swap the batch-index and the time-index of the Neural DE's output
-considering that each time point is a channel.
+We can have Lux's conventional order `(data, channel, batch)` by using it as the last layer
+of `AbstractExplicitLayer` to swap the batch-index and the time-index of the Neural DE's
+output considering that each time point is a channel.
 """
 @concrete struct DimMover <: AbstractExplicitLayer
     from
